check the results of depth-wise convolution and conv2d

+++ backend/opencl/utils/coreUtils.h
@@ -8,7 +8,7 @@
 #include <iostream>
 namespace SNN
 {
-    void getImageShape(const std::vector<uint8_t> &shape, const OpenCLBufferFormat type, size_t *imageShape);
+    void getImageShape(const std::vector<int> &shape, const OpenCLBufferFormat type, size_t *imageShape);
     void CopyBufferToImage(OpenCLRuntime *runtime, const cl_mem &buffer, const cl_mem &image, int *w, int *h, cl_int &err);
     std::vector<int> TensorShapeFormat(const Tensor *input);
 }
